add --debug-env and --java-home switches
