+++ src/Elastic.Configuration/EnvironmentBased/ElasticsearchEnvironmentStateProvider.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.IO;
-using System.Linq;
 using System.Reflection;
 
 namespace Elastic.Configuration.EnvironmentBased
@@ -48,77 +46,4 @@ namespace Elastic.Configuration.EnvironmentBased
 		public void SetEsConfigEnvironmentVariable(string esConfig) =>
 			Environment.SetEnvironmentVariable("ES_CONFIG", esConfig, EnvironmentVariableTarget.Machine);
 	}
-
-	public class ElasticsearchEnvironmentConfiguration
-	{
-		public static ElasticsearchEnvironmentConfiguration Default { get; } = new ElasticsearchEnvironmentConfiguration(new ElasticsearchEnvironmentStateProvider());
-
-		public IElasticsearchEnvironmentStateProvider StateProvider { get; }
-
-		public ElasticsearchEnvironmentConfiguration(IElasticsearchEnvironmentStateProvider stateProvider)
-		{
-			StateProvider = stateProvider ?? new ElasticsearchEnvironmentStateProvider();
-		}
-
-		private string HomeDirectoryInferred
-		{
-			get
-			{
-				var codeBase = StateProvider.RunningExecutableLocation;
-				if (string.IsNullOrWhiteSpace(codeBase)) return null;
-				var codeBaseFolder = new DirectoryInfo(Path.GetDirectoryName(codeBase));
-				var directoryInfo = codeBaseFolder.Parent;
-				return directoryInfo?.FullName;
-			}
-		}
-
-		public string TargetInstallationDirectory => new []
-			{
-				StateProvider.HomeDirectoryProcessVariable,
-				StateProvider.HomeDirectoryUserVariable,
-				StateProvider.HomeDirectoryMachineVariable,
-			}
-			.FirstOrDefault(v=>!string.IsNullOrWhiteSpace(v));
-
-		public string TargetInstallationConfigDirectory => new []
-			{
-				StateProvider.ConfigDirectoryProcessVariable,
-				StateProvider.ConfigDirectoryUserVariable,
-				StateProvider.ConfigDirectoryMachineVariable,
-			}
-			.FirstOrDefault(v=>!string.IsNullOrWhiteSpace(v));
-
-		public string HomeDirectory => new []
-			{
-				StateProvider.HomeDirectoryProcessVariable,
-				StateProvider.HomeDirectoryUserVariable,
-				StateProvider.HomeDirectoryMachineVariable,
-				this.HomeDirectoryInferred
-			}
-			.FirstOrDefault(v=>!string.IsNullOrWhiteSpace(v));
-
-		public string ConfigDirectory
-		{
-			get
-			{
-				var variableOption = new []
-				{
-					StateProvider.ConfigDirectoryProcessVariable,
-					StateProvider.ConfigDirectoryUserVariable,
-					StateProvider.ConfigDirectoryMachineVariable,
-				}.FirstOrDefault(v=>!string.IsNullOrWhiteSpace(v));
-				if (!string.IsNullOrWhiteSpace(variableOption)) return variableOption;
-
-				var homeDir = this.HomeDirectory;
-				return string.IsNullOrEmpty(homeDir) ? null : Path.Combine(homeDir, "config");
-			}
-        }
-
-		public string GetEnvironmentVariable(string variable) => this.StateProvider.GetEnvironmentVariable(variable);
-
-		public void SetEsHomeEnvironmentVariable(string esHome) => StateProvider.SetEsHomeEnvironmentVariable(esHome);
-
-		public void SetEsConfigEnvironmentVariable(string esConfig) => StateProvider.SetEsConfigEnvironmentVariable(esConfig);
-	}
-
 }
